Cast m_pPlayerEntityList.size() to int for EntityManager::playerExists to avoid signed/unsigned mismatch

+++ Source/EntityManager.cpp
@@ -412,7 +412,7 @@ void EntityManager::execute(ePlayer player, eVariableCommand command, float x, f
 
 bool EntityManager::playerExists(ePlayer player)
 {
-    return m_pPlayerEntityList.size() > player;
+    return (int) m_pPlayerEntityList.size() > player;
 }
 
 PlayerEntity* EntityManager::getPlayer(ePlayer player)
